check - more tests
moved all tests out of internal/checker/checker_test.go
moved all these tests into new (modified) internal/repository/check_test.go

modified checkData() from checker_test.go
and renamed it readPackfilesFromRepository(), changed signature.
