+++ internal/checker/checker_test.go
@@ -1,7 +1,6 @@
 package checker_test
 
 import (
-	"bytes"
 	"context"
 	"io"
 	"math/rand"
@@ -568,155 +567,6 @@ func TestCheckerBlobTypeConfusion(t *testing.T) {
 		t.Fatal("no error found, checker is broken")
 	}
 	test.Assert(t, delayRepo.Triggered, "delay repository did not trigger")
-}
-
-// these test were generated by claude.ai, sonnet 4.6
-// --- Helper backend types ---
-
-// lastByteFlipBackend flips the last byte of every pack file on read,
-// causing the SHA-256 hash computed by checkPackInner to differ from the
-// content-addressed pack ID stored in the filename.
-type lastByteFlipBackend struct {
-	backend.Backend
-}
-
-func (b *lastByteFlipBackend) Load(ctx context.Context, h backend.Handle, length int, offset int64, consumer func(rd io.Reader) error) error {
-	if h.Type != restic.PackFile {
-		return b.Backend.Load(ctx, h, length, offset, consumer)
-	}
-	return b.Backend.Load(ctx, h, length, offset, func(rd io.Reader) error {
-		buf, err := io.ReadAll(rd)
-		if err != nil {
-			return err
-		}
-		if len(buf) > 0 {
-			buf[len(buf)-1] ^= 0xff
-		}
-		return consumer(bytes.NewReader(buf))
-	})
-}
-
-// alwaysFailBackend returns a hard, non-partial error for every pack file
-// load, simulating a complete download failure (e.g. network unreachable).
-type alwaysFailBackend struct {
-	backend.Backend
-}
-
-func (b *alwaysFailBackend) Load(ctx context.Context, h backend.Handle, length int, offset int64, consumer func(rd io.Reader) error) error {
-	if h.Type == restic.PackFile {
-		return errors.New("simulated total download failure")
-	}
-	return b.Backend.Load(ctx, h, length, offset, consumer)
-}
-
-// truncatingBackend returns only the first 8 bytes of every pack file,
-// guaranteeing io.ErrUnexpectedEOF inside checkPackInner (a partial read).
-type truncatingBackend struct {
-	backend.Backend
-}
-
-func (b *truncatingBackend) Load(ctx context.Context, h backend.Handle, length int, offset int64, consumer func(rd io.Reader) error) error {
-	if h.Type != restic.PackFile {
-		return b.Backend.Load(ctx, h, length, offset, consumer)
-	}
-	return b.Backend.Load(ctx, h, length, offset, func(rd io.Reader) error {
-		buf := make([]byte, 8)
-		n, _ := io.ReadFull(rd, buf)
-		return consumer(bytes.NewReader(buf[:n]))
-	})
-}
-
-// --- Helper functions ---
-
-// mustLoadSnapshots calls LoadSnapshots and fails the test on error.
-// It must be called before checkData in any test that does not go through
-// checkStruct first, because checker.ReadPacks calls IsFiltered() which
-// dereferences the snapshotFilter pointer that LoadSnapshots initialises.
-func mustLoadSnapshots(t *testing.T, chkr *checker.Checker) {
-	t.Helper()
-	err := chkr.LoadSnapshots(context.TODO(), &data.SnapshotFilter{}, nil)
-	test.OK(t, err)
-}
-
-// setupCheckRepo creates a repository with one snapshot, then re-opens the
-// same backend through the given wrapper for use by the checker.
-// Returning the raw backend separately allows tests to wrap it after data
-// has already been written.
-func setupCheckRepo(t *testing.T, wrap func(backend.Backend) backend.Backend) *checker.Checker {
-	t.Helper()
-	// Write a snapshot into a fresh in-memory repository.
-	repo, be := repository.TestRepositoryWithBackend(t, nil, 0, repository.Options{})
-	_ = archiver.TestSnapshot(t, repo, ".", nil)
-
-	// Re-open the same backend (now containing real pack files) through
-	// the corruption wrapper so the checker reads corrupted data.
-	checkRepo := repository.TestOpenBackend(t, wrap(be))
-	chkr := checker.New(checkRepo, false)
-
-	hints, errs := chkr.LoadIndex(context.TODO(), nil)
-	test.OKs(t, errs)
-	if len(hints) > 0 {
-		assertOnlyMixedPackHints(t, hints)
-	}
-	mustLoadSnapshots(t, chkr)
-	return chkr
-}
-
-// TestCheckPackHashMismatch verifies that checkPackInner detects when the
-// bytes stored in the backend don't hash to the pack's content-addressed ID.
-// Covers the `!hash.Equal(id)` branch → ErrPackData "unexpected pack id".
-func TestCheckPackHashMismatch(t *testing.T) {
-	chkr := setupCheckRepo(t, func(be backend.Backend) backend.Backend {
-		return &lastByteFlipBackend{Backend: be}
-	})
-
-	found := false
-	dataErrs := checkData(chkr)
-	for _, err := range dataErrs {
-		if strings.Contains(err.Error(), "unexpected pack id") {
-			found = true
-		}
-	}
-	test.Assert(t, found, "expected 'unexpected pack id' error, got: %v", dataErrs)
-}
-
-// TestCheckPackDownloadError verifies that a complete (non-partial) backend
-// load failure is returned as a plain "download error" and NOT as ErrPackData.
-// The distinction matters: the check command only suggests `restic repair` for
-// ErrPackData; there is nothing to repair if the file cannot be fetched at all.
-// Covers the non-partialReadError branch of the be.Load error path.
-func TestCheckPackDownloadError(t *testing.T) {
-	chkr := setupCheckRepo(t, func(be backend.Backend) backend.Backend {
-		return &alwaysFailBackend{Backend: be}
-	})
-
-	dataErrs := checkData(chkr)
-	test.Assert(t, len(dataErrs) > 0, "expected download errors, got none")
-	for _, err := range dataErrs {
-		var packErr *repository.ErrPackData
-		test.Assert(t, !errors.As(err, &packErr),
-			"complete download failure must NOT produce ErrPackData, got: %v", err)
-		test.Assert(t, strings.Contains(err.Error(), "download error"),
-			"expected 'download error' in message, got: %v", err)
-	}
-}
-
-// TestCheckPackPartialDownloadError verifies that a partial read (truncated
-// response) is returned as ErrPackData, so the check command can suggest
-// `restic repair packs` for the affected pack.
-// Covers the partialReadError branch of the be.Load error path.
-func TestCheckPackPartialDownloadError(t *testing.T) {
-	chkr := setupCheckRepo(t, func(be backend.Backend) backend.Backend {
-		return &truncatingBackend{Backend: be}
-	})
-
-	dataErrs := checkData(chkr)
-	test.Assert(t, len(dataErrs) > 0, "expected errors from truncated reads, got none")
-	for _, err := range dataErrs {
-		var packErr *repository.ErrPackData
-		test.Assert(t, errors.As(err, &packErr),
-			"partial read must produce ErrPackData, got: %T %v", err, err)
-	}
 }
 
 func loadBenchRepository(t *testing.B) (*checker.Checker, restic.Repository, func()) {
